create new logic in some class
