--- src/AbstractSendable.java
@@ -1,31 +1,0 @@
-public abstract class AbstractSendable implements Sendable {
-    protected final String from;
-    protected final String to;
-
-    public AbstractSendable(String from, String to) {
-        this.from = from;
-        this.to = to;
-    }
-
-    @Override
-    public String getFrom() {
-        return from;
-    }
-
-    @Override
-    public String getTo() {
-        return to;
-    }
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        AbstractSendable that = (AbstractSendable) o;
-
-        if (!from.equals(that.from)) return false;
-        if (!to.equals(that.to)) return false;
-
-        return true;
-    }
-}
