--- src/IllegalPackageException.java
@@ -1,9 +1,0 @@
-
-public  class IllegalPackageException extends RuntimeException{
-    public IllegalPackageException() {
-    }
-
-    public IllegalPackageException(String message) {
-        super(message);
-    }
-}
