--- src/MailService.java
@@ -1,3 +1,0 @@
-interface MailService {
-    Sendable processMail(Sendable mail);
-}
