+++ src/Main.java
@@ -1,40 +1,18 @@
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.logging.*;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 public class Main {
-
-
-
-
-    //org.stepic.java.logging.ClassA
-    private static final Logger LOGGERA = Logger.getLogger(Main.class.getName());
-
-    final Logger LOGGERB = Logger.getLogger("org.stepic.java.logging.ClassB");
-    final Logger LOGGER = Logger.getLogger("org.stepic.java");
-    private static ConsoleHandler handler = new ConsoleHandler();
-    private static  FileHandler fh;
-
-    static {
-        try {
-            fh = new FileHandler("F:\\Temp\\test.txt");
-            fh.setFormatter(new SimpleFormatter());
-            fh.setLevel(Level.WARNING);
-        } catch (IOException e) {
-            //throw new RuntimeException(e);
-        }
-    }
     public static final String AUSTIN_POWERS = "Austin Powers";
     public static final String WEAPONS = "weapons";
     public static final String BANNED_SUBSTANCE = "banned substance";
 
-    public static void main(String[] args){
+    public static void main(String[] args) {
         Logger logger = Logger.getLogger(Main.class.getName());
 
         Inspector inspector = new Inspector();
         Spy spy = new Spy(logger);
         Thief thief = new Thief(10000);
-        MailService variousWorkers[] = new MailService[]{spy, thief, inspector};
+        MailService variousWorkers[] = {spy, thief, inspector};
         UntrustworthyMailWorker worker = new UntrustworthyMailWorker(variousWorkers);
 
         AbstractSendable correspondence[] = {
@@ -52,57 +30,270 @@ public class Main {
                 new MailPackage(AUSTIN_POWERS, "ИГИЛ (запрещенная группировка", new Package("tiny bomb", 9000)),
                 new MailMessage(AUSTIN_POWERS, "Психиатр", "Помогите"),
         };
-        Arrays.stream(correspondence).forEach(parcell -> {
-            try {
-                worker.processMail(parcell);
-            } catch (StolenPackageException e) {
-                logger.log(Level.WARNING, "Inspector found stolen package: " + e);
-            } catch (IllegalPackageException e) {
-                logger.log(Level.WARNING, "Inspector found illegal package: " + e);
-            }
-        });
 
-
-
-    }
-
-
-    private static void configureLogging() throws IOException {
-        final Logger LOGGERA = Logger.getLogger("org.stepic.java.logging.ClassA");
-
-        final Logger LOGGERB = Logger.getLogger("org.stepic.java.logging.ClassB");
-        final Logger LOGGER = Logger.getLogger("org.stepic.java");
-
-        LOGGERA.setLevel(Level.ALL);
-        LOGGERB.setLevel(Level.WARNING);
-
-       ConsoleHandler handler = new ConsoleHandler();
-
-        FileHandler fh = new FileHandler("F:\\Temp\\test.txt");
-        fh.setFormatter(new SimpleFormatter());
-        fh.setLevel(Level.ALL);
-
-        LOGGERA.addHandler(fh);
-
-        handler.setLevel(Level.ALL);
-        handler.setFormatter(new XMLFormatter());
-        LOGGER.addHandler(handler);
-        LOGGER.setUseParentHandlers(false);
-
-
-    }
-
-    public static int[] mergeArrays(int[] a1, int[] a2) {
-        int[] result = new int[a1.length + a2.length];
-        int pos1 = 0;
-        int pos2 = 0;
-        while (pos1 < a1.length || pos2 < a2.length) {
-            result[pos1 + pos2] = (pos1 < a1.length && (pos2 == a2.length || a1[pos1] < a2[pos2]) ?
-                    a1[pos1++] : a2[pos2++]);
+        for (AbstractSendable as : correspondence) {
+            worker.processMail(as);
         }
-        return result;
 
 
-        // return sortedArray; // your implementation here
     }
+
+    static interface Sendable {
+        String getFrom();
+
+        String getTo();
+    }
+
+    public static abstract class AbstractSendable implements Sendable {
+        protected final String from;
+        protected final String to;
+
+        public AbstractSendable(String from, String to) {
+            this.from = from;
+            this.to = to;
+        }
+
+        @Override
+        public String getFrom() {
+            return from;
+        }
+
+        @Override
+        public String getTo() {
+            return to;
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+
+            AbstractSendable that = (AbstractSendable) o;
+
+            if (!from.equals(that.from)) return false;
+            if (!to.equals(that.to)) return false;
+
+            return true;
+        }
+    }
+
+    public static class MailMessage extends AbstractSendable {
+        private final String message;
+
+        public MailMessage(String from, String to, String message) {
+            super(from, to);
+            this.message = message;
+        }
+
+        public String getMessage() {
+            return message;
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+            if (!super.equals(o)) return false;
+
+            MailMessage that = (MailMessage) o;
+
+            if (message != null ? !message.equals(that.message) : that.message != null) return false;
+
+            return true;
+        }
+    }
+
+    public static class MailPackage extends AbstractSendable {
+        private final Package content;
+
+        public MailPackage(String from, String to, Package con) {
+            super(from, to);
+            content = con;
+
+        }
+
+
+        public Package getContent() {
+            return content;
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+            if (!super.equals(o)) return false;
+
+            MailPackage that = (MailPackage) o;
+
+            if (!content.equals(that.content)) return false;
+
+            return true;
+        }
+    }
+
+    public static class Package {
+        private static String content;
+        private final int price;
+
+        public Package(String content, int price) {
+            this.content = content;
+            this.price = price;
+        }
+
+        public String getContent() {
+            return content;
+        }
+
+        public int getPrice() {
+            return price;
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) return true;
+            if (o == null || getClass() != o.getClass()) return false;
+
+            Package aPackage = (Package) o;
+
+            if (price != aPackage.price) return false;
+            if (!content.equals(aPackage.content)) return false;
+
+            return true;
+        }
+    }
+
+    public static class RealMailService implements MailService {
+        @Override
+        public Sendable processMail(Sendable mail) {
+
+            return mail;
+        }
+    }
+
+
+    public static class UntrustworthyMailWorker implements MailService {
+        RealMailService realMailService = new RealMailService();
+        private Logger loger = Logger.getLogger(UntrustworthyMailWorker.class.getName());
+        Spy spy = new Spy(loger);
+        Thief thief = new Thief(100);
+        Inspector inspector = new Inspector();
+
+        MailService[] ms = {spy, inspector, thief};
+
+        public UntrustworthyMailWorker(MailService[] mailServices) {
+            ms = mailServices;
+        }
+
+        RealMailService getRealMailService() {
+            return realMailService;
+        }
+
+        @Override
+        public Sendable processMail(Sendable mail) {
+            if (mail instanceof Sendable) {
+                spy.processMail(mail);
+                thief.processMail(mail);
+                inspector.processMail(mail);
+                realMailService.processMail(mail);
+            }
+            return mail;
+        }
+    }
+
+    public static class Spy implements MailService {
+        Logger logger;
+        String fromName = null;
+        String toName = null;
+
+        public Spy(Logger logger) {
+            this.logger = logger;
+        }
+
+        MailMessage mailMessage = new MailMessage("from", "to", "someMessage");
+
+
+        @Override
+        public Sendable processMail(Sendable mail) {
+            if (mail instanceof MailMessage) {
+                if (((MailMessage) mail).from.equals(AUSTIN_POWERS)) {
+                    logger.log(Level.WARNING, "Detected target mail correspondence: from {0} to {1} \"{2}\"", new String[]{((MailMessage) mail).from, ((MailMessage) mail).to, ((MailMessage) mail).message});
+                } else {
+                    logger.log(Level.INFO, "Usual correspondence: from {0} to {1}", new String[]{((MailMessage) mail).from, ((MailMessage) mail).to});
+                }
+            }
+            return mail;
+        }
+    }
+
+    public static class Thief implements MailService {
+        private static int minPricePackage;
+        private static int stolenPrice;
+        String cont = null;
+
+
+        public Thief(int minPrice) {
+            minPricePackage = minPrice;
+        }
+
+        int getStolenValue(int costPack) {
+            stolenPrice = stolenPrice + costPack;
+            return stolenPrice;
+        }
+
+
+        @Override
+        public Sendable processMail(Sendable mail) {
+
+            if (mail instanceof Package) {
+                if (((Package) mail).getPrice() > minPricePackage) {
+                    getStolenValue(((Package) mail).getPrice());
+                    cont = ((Package) mail).getContent();
+                    Package p = new Package("stones instead of " + cont, 0);
+                    mail = (Sendable) p;
+                }
+            }
+            return mail;
+        }
+    }
+
+    public static class Inspector implements MailService {
+
+        @Override
+        public Sendable processMail(Sendable mail) {
+            if (mail instanceof Package) {
+                if (((Package) mail).getContent().matches("weapons") || ((Package) mail).getContent().matches("banned substance")) {
+                    throw new IllegalPackageException();
+                }
+                if (((Package) mail).getContent().matches("stones")) {
+                    throw new StolenPackageException();
+                }
+            }
+            return mail;
+        }
+    }
+
+    public static class StolenPackageException extends RuntimeException {
+        public StolenPackageException() {
+        }
+
+        public StolenPackageException(String message) {
+            super(message);
+        }
+    }
+
+    public static class IllegalPackageException extends RuntimeException {
+        public IllegalPackageException() {
+        }
+
+        public IllegalPackageException(String message) {
+            super(message);
+        }
+    }
+
+
+    static interface MailService {
+        Sendable processMail(Sendable mail);
+    }
+
+
 }
