--- src/Package.java
@@ -1,30 +1,0 @@
-public class Package {
-    private static String content;
-    private final int price;
-
-    public Package(String content, int price) {
-        this.content = content;
-        this.price = price;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        Package aPackage = (Package) o;
-
-        if (price != aPackage.price) return false;
-        if (!content.equals(aPackage.content)) return false;
-
-        return true;
-    }
-}
