--- src/RealMailService.java
@@ -1,7 +1,0 @@
-public class    RealMailService implements MailService{
-    @Override
-    public Sendable processMail(Sendable mail) {
-
-        return mail;
-    }
-}
