--- src/Sendable.java
@@ -1,5 +1,0 @@
-interface Sendable {
-    String getFrom();
-
-    String getTo();
-}
