--- src/StolenPackageException.java
@@ -1,8 +1,0 @@
-public  class StolenPackageException extends RuntimeException{
-    public StolenPackageException() {
-    }
-
-    public StolenPackageException(String message) {
-        super(message);
-    }
-}
